Fix leak
